+++ apps/api/CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   `getResolveObjects` method.
+
 ### Changed
 
 ### Deprecated
+
+-   `postResolveObjects` method.
 
 ### Fixed