+++ apps/www/CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Image List View on Home Page to Image Rail.
 -   Refactored styling for content width.
 -   Hiding drag-and-drop prompt for Collections Drawer if device does not have hovering capabilities.
+-   Using `GET /resolve/{authority}/{namespace}` instead of `POST /resolve/{authority}/{namespace}` for external resolvers.
+-   Updated examples for resolving external objects in the _API Recipes_ article.
 
 ### Deprecated